ATLAS-52 Add notification for create element and allerts for delete

+++ src/admin/components/AnatomicalStructureList/index.tsx
@@ -1,43 +1,31 @@
 import { useEffect, useState } from "react";
 import TableComponent from "../../../components/UI/TableComponent";
-import { deleteAnatomicalStructure, getAnatomicalStructureList } from "../../../requests/anatomicalStructureRequests";
+import { deleteAnatomicalStructure } from "../../../requests/anatomicalStructureRequests";
 import { AnatomicalStructure } from "../../../_types";
 
-const AnatomicalStructureList = ({ anatomicalStructureList }) => {
+type Props = {
+    anatomicalStructureList: AnatomicalStructure[];
+    columns: string[];
+};
+
+const AnatomicalStructureList = ({ anatomicalStructureList, columns }: Props) => {
     const [structureList, setStructureList] = useState<AnatomicalStructure[]>(anatomicalStructureList);
-    const [columns, setColumns] = useState<string[]>([]);
 
     useEffect(() => {
-        const fetchData = async () => {
-            try {
-                const result = await getAnatomicalStructureList();
-                setStructureList(result);
-            } catch (error) {
-                console.error("Error fetching AnatomicalStructureList:", error);
-            }
-        };
-
-        if (!structureList) {
-            fetchData();
-        }
-    }, [structureList]);
-
-    useEffect(() => {
-        if (structureList.length) {
-            const columnsTitles = Object.keys(structureList[0]);
-            columnsTitles.push("Actions");
-            setColumns(columnsTitles);
-        }
-    }, [structureList]);
+        setStructureList(anatomicalStructureList);
+    }, [anatomicalStructureList]);
 
     const removeItemById = async (itemId: string) => {
-        try {
-            const result = await deleteAnatomicalStructure(itemId);
-            if (result === 204) {
-                setStructureList(structureList.filter((item) => item.id !== itemId));
+        const isAlert = confirm("уверены что хотите удалить?");
+        if (isAlert) {
+            try {
+                const result = await deleteAnatomicalStructure(itemId);
+                if (result === 204) {
+                    setStructureList(structureList.filter((item) => item.id !== itemId));
+                }
+            } catch (error) {
+                console.error("Error fetching AnatomicalStructureSubjectList:", error);
             }
-        } catch (error) {
-            console.error("Error fetching AnatomicalStructureSubjectList:", error);
         }
     };
