+++ src/admin/components/StudySeriesList/index.tsx
@@ -24,13 +24,16 @@ const StudySeriesList = ({ seriesId }) => {
     }, []);
 
     const removeItemById = async (itemId: string) => {
-        try {
-            const result = await deleteStudySeries(itemId);
-            if (result === 204) {
-                setStudySeriesList(studySeriesList.filter((item) => item.id !== itemId));
+        const isAlert = confirm("уверены что хотите удалить?");
+        if (isAlert) {
+            try {
+                const result = await deleteStudySeries(itemId);
+                if (result === 204) {
+                    setStudySeriesList(studySeriesList.filter((item) => item.id !== itemId));
+                }
+            } catch (error) {
+                console.error("Error fetching StudySeriesList:", error);
             }
-        } catch (error) {
-            console.error("Error fetching StudySeriesList:", error);
         }
     };
